Make the new behavior the default, for the unwrapNode
